backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
rembg==2.0.67
Pillow==10.4.0
onnxruntime==1.18.1
numpy==1.26.4
